refactor(ios): split notification target index types
