+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/NotificationFeedWorkspaceLookupKey.swift
@@ -1,0 +1,4 @@
+struct NotificationFeedWorkspaceLookupKey: Hashable, Sendable {
+    let macDeviceID: String
+    let targetID: String
+}
