+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/NotificationFeedWorkspaceTargetIndex.swift
@@ -2,68 +2,29 @@ import CmuxMobileShellModel
 
 /// Resolves notification destinations from one immutable workspace snapshot.
 struct NotificationFeedWorkspaceTargetIndex: Sendable {
-    private struct LookupKey: Hashable, Sendable {
-        let macDeviceID: String
-        let targetID: String
-    }
-
-    private struct Target: Sendable {
-        var rowIDs: Set<MobileWorkspacePreview.ID> = []
-        var exactRowIDsByInstanceTag: [String: Set<MobileWorkspacePreview.ID>] = [:]
-        var owners: Set<MacPairingKey> = []
-        var ownerDevices: Set<String> = []
-
-        mutating func insert(
-            rowID: MobileWorkspacePreview.ID,
-            macDeviceID: String,
-            instanceTag: String?
-        ) {
-            rowIDs.insert(rowID)
-            if let instanceTag, !instanceTag.isEmpty {
-                exactRowIDsByInstanceTag[instanceTag, default: []].insert(rowID)
-            }
-            let owner = MacPairingKey(
-                macDeviceID: macDeviceID,
-                instanceTag: instanceTag
-            )
-            owners.insert(owner)
-            ownerDevices.insert(owner.canonicalMacDeviceID)
-        }
-
-        func rowID(instanceTag: String?) -> MobileWorkspacePreview.ID? {
-            if let instanceTag, !instanceTag.isEmpty {
-                guard let exactRowIDs = exactRowIDsByInstanceTag[instanceTag],
-                      exactRowIDs.count == 1 else { return nil }
-                return exactRowIDs.first
-            }
-            guard owners.count <= ownerDevices.count, rowIDs.count == 1 else { return nil }
-            return rowIDs.first
-        }
-    }
-
-    private var workspacesByRemoteID: [LookupKey: Target] = [:]
-    private var workspacesBySurfaceID: [LookupKey: Target] = [:]
+    private var workspacesByRemoteID: [NotificationFeedWorkspaceLookupKey: NotificationFeedWorkspaceTarget] = [:]
+    private var workspacesBySurfaceID: [NotificationFeedWorkspaceLookupKey: NotificationFeedWorkspaceTarget] = [:]
 
     init(workspaces: [MobileWorkspacePreview]) {
         for workspace in workspaces {
             guard let macDeviceID = workspace.macDeviceID, !macDeviceID.isEmpty else {
                 continue
             }
-            let workspaceKey = LookupKey(
+            let workspaceKey = NotificationFeedWorkspaceLookupKey(
                 macDeviceID: macDeviceID,
                 targetID: workspace.rpcWorkspaceID.rawValue
             )
-            workspacesByRemoteID[workspaceKey, default: Target()].insert(
+            workspacesByRemoteID[workspaceKey, default: NotificationFeedWorkspaceTarget()].insert(
                 rowID: workspace.id,
                 macDeviceID: macDeviceID,
                 instanceTag: workspace.macInstanceTag
             )
             for terminal in workspace.terminals {
-                let surfaceKey = LookupKey(
+                let surfaceKey = NotificationFeedWorkspaceLookupKey(
                     macDeviceID: macDeviceID,
                     targetID: terminal.id.rawValue
                 )
-                workspacesBySurfaceID[surfaceKey, default: Target()].insert(
+                workspacesBySurfaceID[surfaceKey, default: NotificationFeedWorkspaceTarget()].insert(
                     rowID: workspace.id,
                     macDeviceID: macDeviceID,
                     instanceTag: workspace.macInstanceTag
@@ -76,7 +37,7 @@ struct NotificationFeedWorkspaceTargetIndex: Sendable {
         for item: MobileNotificationFeedItem
     ) -> MobileWorkspacePreview.ID? {
         let targetID: String
-        let targets: [LookupKey: Target]
+        let targets: [NotificationFeedWorkspaceLookupKey: NotificationFeedWorkspaceTarget]
         if item.retargetsToLiveSurfaceOwner, let surfaceID = item.remoteSurfaceID {
             targetID = surfaceID
             targets = workspacesBySurfaceID
@@ -85,7 +46,7 @@ struct NotificationFeedWorkspaceTargetIndex: Sendable {
             targets = workspacesByRemoteID
         }
         return targets[
-            LookupKey(macDeviceID: item.macDeviceID, targetID: targetID)
+            NotificationFeedWorkspaceLookupKey(macDeviceID: item.macDeviceID, targetID: targetID)
         ]?.rowID(instanceTag: item.macInstanceTag)
     }
 }
